Correct MKMCEngine and used it in the test (test pass)
