+++ test/Makefile
@@ -1,9 +1,6 @@
 CXXFLAGS=-Wall
 
-all : equityoptiontest blackscholesconsttest
+all : equityoptiontest 
 
-equityoptiontest : equityoptiontest.cpp
-	g++ -g -o equityoptiontest equityoptiontest.cpp -l QuantLib
-
-blackscholesconsttest : ../src/blackscholesconstprocess.cpp blackscholesconsttest.cpp 
-	g++ -g -o blackscholesconsttest ../src/blackscholesconstprocess.cpp blackscholesconsttest.cpp -l QuantLib
+equityoptiontest : ../src/blackscholesconstprocess.cpp equityoptiontest.cpp 
+	g++ -g -o equityoptiontest ../src/blackscholesconstprocess.cpp equityoptiontest.cpp -l QuantLib
